slam-rs: keep standalone reflection probes in private QR tests
