+++ packages/slam-rs/crates/slam-rs/src/qr.rs
@@ -46,7 +46,9 @@ pub(crate) fn make_givens<S: LieScalar>(p: S, q: S) -> GivensRotation<S> {
 
 #[cfg(test)]
 mod tests {
+    #![allow(clippy::unwrap_used, clippy::expect_used)]
     use super::*;
+    use nalgebra::DVector;
     use proptest::prelude::*;
 
     proptest! {
@@ -84,5 +86,125 @@ mod tests {
             prop_assert!(a[(1, 0)].abs() < 1e-12);
             prop_assert!((a.norm() - p.hypot(q)).abs() < 1e-12);
         }
+    }
+
+    // Test-only driver; production callers own their scratch and validated views.
+    fn reflect_column<S: LieScalar>(
+        storage: &mut DMatrix<S>,
+        col: usize,
+        start: usize,
+        len: usize,
+    ) {
+        let mut axis = vec![S::zero(); len];
+        let (active, _) = make_householder(storage, col, start, len, &mut axis);
+        let cols = storage.ncols();
+        apply_householder_on_the_left(storage.view_mut((start, 0), (len, cols)), &axis, active);
+    }
+
+    proptest! {
+        #[test]
+        fn reflections_preserve_norms_and_form_an_orthogonal_q(
+            values in prop::collection::vec(-10.0f64..10.0, 24),
+        ) {
+            let a = DMatrix::from_row_slice(6, 4, &values);
+            // Carry an identity beside A to observe the same sequence of reflections.
+            let mut augmented = DMatrix::zeros(6, 10);
+            augmented.columns_mut(0, 4).copy_from(&a);
+            augmented.columns_mut(4, 6).copy_from(&DMatrix::identity(6, 6));
+            for col in 0..4 {
+                reflect_column(&mut augmented, col, col, 6 - col);
+            }
+            let r = augmented.columns(0, 4);
+            let qt = augmented.columns(4, 6);
+            prop_assert!((qt * qt.transpose() - DMatrix::identity(6, 6)).norm() < 1e-12);
+            prop_assert!((r.norm() - a.norm()).abs() < 1e-12 * (1.0 + a.norm()));
+            prop_assert!((qt * &a - r).norm() < 1e-12 * (1.0 + a.norm()));
+            for col in 0..4 {
+                for row in col + 1..6 {
+                    prop_assert!(r[(row, col)].abs() < 1e-12 * (1.0 + a.norm()));
+                }
+            }
+        }
+    }
+
+    /// `QRvsLLT` and `QRvsLLTRankDef`, as the assertion they
+    /// demonstrate: for a `10 x 6` `J`, the `R` of a Householder QR and the `Lᵀ` of
+    /// the Cholesky of `JᵀJ` agree row by row up to a sign.
+    ///
+    /// The QR is built from this port's own crate Householder primitives, so
+    /// the test is on the port, not on nalgebra.
+    #[test]
+    fn householder_qr_matches_the_cholesky_of_the_normal_equations() {
+        for (case, rank_deficient) in [("full rank", false), ("rank deficient", true)] {
+            let mut seed: u64 = 0xfeed_1235;
+            let mut j: DMatrix<f64> = DMatrix::zeros(10, 6);
+            for r in 0..10 {
+                for c in 0..6 {
+                    seed ^= seed >> 12;
+                    seed ^= seed << 25;
+                    seed ^= seed >> 27;
+                    let value = seed.wrapping_mul(0x2545_f491_4f6c_dd1d);
+                    j[(r, c)] = (value >> 11) as f64 / (1u64 << 53) as f64 * 2.0 - 1.0;
+                }
+            }
+            if rank_deficient {
+                // `J.col(2) = J.col(4)`.
+                let col4: DVector<f64> = j.column(4).into_owned();
+                j.set_column(2, &col4);
+            }
+
+            // The port has no standalone QR — the landmark block eliminates exactly
+            // three columns — so the reflections are driven here the way
+            // `performQRHouseholder` drives them.
+            let r_qr: DMatrix<f64> = householder_r(&j);
+            let ata: DMatrix<f64> = j.transpose() * &j;
+
+            if rank_deficient {
+                // For singular `JᵀJ`, QR still yields triangular `R` with `RᵀR = JᵀJ`.
+                let rtr: DMatrix<f64> = r_qr.transpose() * &r_qr;
+                assert!(
+                    (&rtr - &ata).norm() <= 1e-9 * ata.norm(),
+                    "{case}: RᵀR != JᵀJ"
+                );
+                continue;
+            }
+
+            let llt = ata.clone().cholesky().expect("full rank");
+            let u: DMatrix<f64> = llt.l().transpose();
+            for row in 0..6 {
+                // Up to the sign of the row, which is all a QR is free in.
+                let sign: f64 = if r_qr[(row, row)] * u[(row, row)] < 0.0 {
+                    -1.0
+                } else {
+                    1.0
+                };
+                for col in 0..6 {
+                    let got: f64 = sign * r_qr[(row, col)];
+                    assert!(
+                        (got - u[(row, col)]).abs() <= 1e-9 * u.norm(),
+                        "{case}: R[{row},{col}] = {got}, Lᵀ = {}",
+                        u[(row, col)]
+                    );
+                }
+            }
+        }
+    }
+
+    /// The upper-triangular factor of a Householder QR, driven exactly as
+    /// `performQRHouseholder` drives it, through the port's own primitives.
+    fn householder_r(j: &DMatrix<f64>) -> DMatrix<f64> {
+        let rows: usize = j.nrows();
+        let cols: usize = j.ncols();
+        let mut work: DMatrix<f64> = j.clone();
+        for k in 0..cols {
+            reflect_column(&mut work, k, k, rows - k);
+        }
+        let mut r: DMatrix<f64> = DMatrix::zeros(cols, cols);
+        for row in 0..cols {
+            for col in row..cols {
+                r[(row, col)] = work[(row, col)];
+            }
+        }
+        r
     }
 }
